requirements.txt
anthropic>=0.18.0
praw>=7.7.1
beautifulsoup4>=4.12.0
requests>=2.31.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
selenium>=4.15.0
webdriver-manager>=4.0.1
pandas>=2.1.0
pyyaml>=6.0.1
rich>=13.7.0
tenacity>=8.2.3
aiohttp>=3.9.0
asyncio>=3.4.3
